requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
tkcalendar==1.6.1
pyinstaller==6.3.0  # if you want to build your app
ttkthemes==3.2.2
customtkinter==5.2.2
Pillow==10.2.0
CTkMessagebox==2.5
